week-1/requirements.txt
asyncio
braintrust
instructor
kagglehub
openai
pydantic
python-dotenv
rich
tenacity